backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "langgraph>=0.2.0",
    "pdfplumber>=0.10.0",
    "chromadb>=0.4.0",
    "langchain-chroma>=0.1.0",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=2.2.0",
    "tiktoken>=0.7.0",
]
